# Replace `encoding/json` with `json-iterator/go` in backend