@zalegrala 2024-10-18
```
❯ benchstat main.txt jsonIter.txt
goos: linux
goarch: amd64
pkg: github.com/grafana/tempo/tempodb/backend
cpu: AMD Ryzen 7 PRO 7840U w/ Radeon 780M Graphics
                  │  main.txt   │            jsonIter.txt             │
                  │   sec/op    │   sec/op     vs base                │
IndexUnmarshal-16   57.49µ ± 3%   41.67µ ± 2%  -27.52% (p=0.000 n=10)

                  │   main.txt   │            jsonIter.txt             │
                  │     B/op     │     B/op      vs base               │
IndexUnmarshal-16   55.71Ki ± 0%   53.02Ki ± 0%  -4.82% (p=0.000 n=10)

                  │  main.txt  │            jsonIter.txt             │
                  │ allocs/op  │ allocs/op   vs base                 │
IndexUnmarshal-16   206.0 ± 1%   421.0 ± 0%  +104.37% (p=0.000 n=10)

```